+++ bids-validator/validators/ometiff/index.js
@@ -1,0 +1,11 @@
+import ometiff from './ometiff'
+import validate from './validate'
+import checkSamples from './checkSamples'
+import checkJSONAndField from './checkJSONAndField'
+
+export default {
+  ometiff,
+  validate,
+  checkSamples,
+  checkJSONAndField,
+}
